feat: add CRUD operations for study topics in StudyTopicController and StudyTopicService

+++ src/main/kotlin/com/github/Zambrinn/KotlinAPI/service/StudyTopicService.kt
@@ -30,19 +30,30 @@ class StudyTopicService(
 
     fun getTopicById(id: Long): StudyTopicResponse? {
         val entityFound = repository.findByIdOrNull(id);
-
         return entityFound?.toResponse();
     }
 
-    fun deleteTopicById(id: Long): StudyTopicResponse? {
-        val entityToDelete = repository.deleteById(id)
-
-        return entityToDelete?.toResponse();
+    fun updateTopic(id: Long, request: StudyTopicRequest): StudyTopicResponse? {
+        return repository.findByIdOrNull(id)?.let { existingTopic ->
+            val updatedTopic = existingTopic.copy(
+                name = request.name,
+                category = request.category
+            )
+            repository.save(updatedTopic).toResponse()
+        }
     }
+    fun deleteTopic(id: Long): Boolean {
+        if (repository.existsById(id)) {
+            repository.deleteById(id)
+            return true
+        }
+        return false
+    }
+
 
     private fun StudyTopic.toResponse(): StudyTopicResponse {
         return StudyTopicResponse(
-            id = this.id!!,
+            id = this.id!!, // non nullable object
             name = this.name,
             category = this.category,
             status = this.status,
